refactor(setup): extract wizard step 5 (config modification)
